refactor(date-picker): move reusable code into utils
